+++ src/main/java/org/example/wishlist/repository/UserRepository.java
@@ -1,0 +1,22 @@
+package org.example.wishlist.repository;
+
+
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.datasource.DriverManagerDataSource;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public class UserRepository {
+
+
+    public UserRepository() {
+        DriverManagerDataSource dataSource = new DriverManagerDataSource(
+                System.getenv("DB_URL"),
+                System.getenv("DB_USERNAME"),
+                System.getenv("DB_PASSWORD")
+        );
+        dataSource.setDriverClassName("com.mysql.cj.jdbc.Driver");
+        JdbcTemplate jdbcTemplate = new JdbcTemplate(dataSource);
+    }
+
+}
