+++ src/main/java/org/example/wishlist/repository/WishRepository.java
@@ -2,7 +2,6 @@ package org.example.wishlist.repository;
 
 import org.example.wishlist.model.Wish;
 import org.example.wishlist.model.Wishlist;
-import org.example.wishlist.model.User;
 import org.example.wishlist.model.WishListRowMapper;
 import org.example.wishlist.model.WishRowMapper;
 import org.springframework.jdbc.core.JdbcTemplate;
@@ -31,10 +30,6 @@ public class WishRepository {
         return jdbcTemplate.query(sql, new WishRowMapper());
     }
 
-    public List<Wishlist> getAllWishList() {
-        String sql = "SELECT name, description FROM wishlist";
-        return jdbcTemplate.query(sql, new WishListRowMapper());
-    }
 
     public void addWish(Wish wish) {
         String sql = "INSERT INTO wish(name, description, price, link) VALUES (?,?,?,?)";
@@ -46,10 +41,7 @@ public class WishRepository {
         jdbcTemplate.update(sql, wishlistId, wish.getWishName(), wish.getDescription(), wish.getLink(), wish.getPrice());
     } //TODO SKAL RYKKES TIL WISHLIST WISHLIST
 
-    public void addWishlist(Wishlist wishlist, int userID) {
-        String sql = "INSERT INTO WISHLIST (USER_ID, NAME, DESCRIPTION) VALUES (?,?,?)";
-        jdbcTemplate.update(sql, userID, wishlist.getWishlistName(), wishlist.getWishlistDesc());
-    }
+
 
     public String deleteWish(int wishID) {
         String sql = "DELETE FROM wish WHERE WISH_ID = ?";
@@ -57,21 +49,11 @@ public class WishRepository {
         return "delete success";
     }
 
-    public String deleteWishlist(int wishlistID) {
-        String sql = "DELETE FROM wishlist WHERE wishlist_ID = ?";
-        jdbcTemplate.update(sql, wishlistID);
-        return "delete success";
-    }
 
     public void updateWish(Wish wish) {
         String sql = "UPDATE wish SET wishName = ?, description = ?, price = ?, link = ? WHERE id = ?";
         jdbcTemplate.update(sql, wish.getWishName(), wish.getDescription(), wish.getPrice(), wish.getLink()); //wish.getId()) burde den ikke være der?
     }
 
-    public void updateWishlist(String name, Wishlist newWishlist) { //String id i stedet for name?
-        String sql = "UPDATE wishlist SET wish = ?, wishlist = ?, price = ?, wishlistDesc = ?, wishlistName = ? WHERE id = ?";
-        jdbcTemplate.update(sql, newWishlist.getWishlist(), newWishlist.getPrice(), newWishlist.getWishlistDesc(),
-                newWishlist.getWishlistName()); //Skal der ikke være et ID her?? //Hvorfor er der et wish i wishlist og en liste af wishes?
 
-    }
 }
